+++ CHANGELOG.md
@@ -24,6 +24,12 @@ Web:
 
 * Add a new hidden feature to set nginx alias in www/custom.yaml.
 
+MTA-STS:
+
+* Added support for client side MTA-STS when there is a valid SSL Certificate on the primary domain
+* Automatically adds reporting when alias "tlsrpt@<primary-domain>" is added.
+* Starts default on 'testing', but changes will be kept between MiaB Upgrades.
+
 Setup:
 
 * Improved error handling.